+++ src/game/Object/Player.cpp
@@ -4571,157 +4571,6 @@ void Player::UpdateZone(uint32 newZone, uint32 newArea)
     UpdateZoneDependentPets();
 }
 
-// If players are too far way of duel flag... then player loose the duel
-void Player::CheckDuelDistance(time_t currTime)
-{
-    if (!duel)
-    {
-        return;
-    }
-
-    GameObject* obj = GetMap()->GetGameObject(GetGuidValue(PLAYER_DUEL_ARBITER));
-    if (!obj)
-    {
-        // player not at duel start map
-        DuelComplete(DUEL_FLED);
-        return;
-    }
-
-    if (duel->outOfBound == 0)
-    {
-        if (!IsWithinDistInMap(obj, 50))
-        {
-            duel->outOfBound = currTime;
-
-            WorldPacket data(SMSG_DUEL_OUTOFBOUNDS, 0);
-            GetSession()->SendPacket(&data);
-        }
-    }
-    else
-    {
-        if (IsWithinDistInMap(obj, 40))
-        {
-            duel->outOfBound = 0;
-
-            WorldPacket data(SMSG_DUEL_INBOUNDS, 0);
-            GetSession()->SendPacket(&data);
-        }
-        else if (currTime >= (duel->outOfBound + 10))
-        {
-            DuelComplete(DUEL_FLED);
-        }
-    }
-}
-
-/**
- * @brief Finalizes an active duel and cleans up all duel state.
- *
- * @param type The duel completion result.
- */
-void Player::DuelComplete(DuelCompleteType type)
-{
-    // duel not requested
-    if (!duel)
-    {
-        return;
-    }
-
-    WorldPacket data(SMSG_DUEL_COMPLETE, (1));
-    data << (uint8)((type != DUEL_INTERRUPTED) ? 1 : 0);
-    GetSession()->SendPacket(&data);
-
-    if (duel->opponent->GetSession())
-    {
-        duel->opponent->GetSession()->SendPacket(&data);
-    }
-
-    if (type != DUEL_INTERRUPTED)
-    {
-        data.Initialize(SMSG_DUEL_WINNER, (1 + 20));          // we guess size
-        data << (uint8)((type == DUEL_WON) ? 0 : 1);          // 0 = just won; 1 = fled
-        data << duel->opponent->GetName();
-        data << GetName();
-        SendMessageToSet(&data, true);
-    }
-
-    // Used by Eluna
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = GetEluna())
-    {
-        e->OnDuelEnd(duel->opponent, this, type);
-    }
-#endif /* ENABLE_ELUNA */
-
-    // Remove Duel Flag object
-    if (GameObject* obj = GetMap()->GetGameObject(GetGuidValue(PLAYER_DUEL_ARBITER)))
-    {
-        duel->initiator->RemoveGameObject(obj, true);
-    }
-
-    /* remove auras */
-    // TODO: Needs a simpler method
-    std::vector<uint32> auras2remove;
-    SpellAuraHolderMap const& vAuras = duel->opponent->GetSpellAuraHolderMap();
-    for (SpellAuraHolderMap::const_iterator i = vAuras.begin(); i != vAuras.end(); ++i)
-    {
-        SpellAuraHolder const* aura = i->second;
-        if (!aura->IsPositive() && aura->GetCasterGuid() == GetObjectGuid() && aura->GetAuraApplyTime() >= duel->startTime)
-        {
-            auras2remove.push_back(aura->GetId());
-        }
-    }
-
-    for (size_t i = 0; i < auras2remove.size(); ++i)
-    {
-        duel->opponent->RemoveAurasDueToSpell(auras2remove[i]);
-    }
-
-    auras2remove.clear();
-    SpellAuraHolderMap const& auras = GetSpellAuraHolderMap();
-    for (SpellAuraHolderMap::const_iterator i = auras.begin(); i != auras.end(); ++i)
-    {
-        SpellAuraHolder const* aura = i->second;
-        if (!aura->IsPositive() && aura->GetCasterGuid() == duel->opponent->GetObjectGuid() && aura->GetAuraApplyTime() >= duel->startTime)
-        {
-            auras2remove.push_back(aura->GetId());
-        }
-    }
-    for (size_t i = 0; i < auras2remove.size(); ++i)
-    {
-        RemoveAurasDueToSpell(auras2remove[i]);
-    }
-
-    // cleanup combo points
-    if (GetComboTargetGuid() == duel->opponent->GetObjectGuid())
-    {
-        ClearComboPoints();
-    }
-    else if (GetComboTargetGuid() == duel->opponent->GetPetGuid())
-    {
-        ClearComboPoints();
-    }
-
-    if (duel->opponent->GetComboTargetGuid() == GetObjectGuid())
-    {
-        duel->opponent->ClearComboPoints();
-    }
-    else if (duel->opponent->GetComboTargetGuid() == GetPetGuid())
-    {
-        duel->opponent->ClearComboPoints();
-    }
-
-    // cleanups
-    SetGuidValue(PLAYER_DUEL_ARBITER, ObjectGuid());
-    SetUInt32Value(PLAYER_DUEL_TEAM, 0);
-    duel->opponent->SetGuidValue(PLAYER_DUEL_ARBITER, ObjectGuid());
-    duel->opponent->SetUInt32Value(PLAYER_DUEL_TEAM, 0);
-
-    delete duel->opponent->duel;
-    duel->opponent->duel = NULL;
-    delete duel;
-    duel = NULL;
-}
-
 /**
  * @brief Sends a single world state update to the client.
  *
@@ -8960,18 +8809,6 @@ void Player::SetRestType(RestType n_r_type, uint32 areaTriggerId /*= 0*/)
             SetFFAPvP(false);
         }
     }
-}
-
-/**
- * @brief Sends the duel countdown timer to the client.
- *
- * @param counter The countdown value in seconds.
- */
-void Player::SendDuelCountdown(uint32 counter)
-{
-    WorldPacket data(SMSG_DUEL_COUNTDOWN, 4);
-    data << uint32(counter);                                // seconds
-    GetSession()->SendPacket(&data);
 }
 
 /**
